pyproject.toml
[project]
name = "xiaomiimeicheckertelegrambot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aiohttp>=3.12.15",
    "python-dotenv>=1.1.1",
]
